feat: move uiState components to uiComponents module
