+++ features/albums/create/src/main/kotlin/features/albums/create/presentation/steps/CreateAlbumScreen.kt
@@ -36,7 +36,7 @@ import platform.uicomponents.components.errorview.DefaultErrorView
 import platform.uicomponents.components.errorview.DefaultErrorViewOptions
 import platform.uicomponents.components.spacers.VerticalSpacer
 import platform.uicomponents.components.spacers.WeightSpacer
-import platform.uistate.uievent.UiEventEffect
+import platform.uicomponents.sideeffects.UiEventEffect
 import platform.uistate.uistate.UiStateContent
 
 internal data class CreateAlbumScreen(
